SedInstance: create a SedInstanceTask object for each task.
