+++ src/api/libopencor/sedtask.h
@@ -28,7 +28,7 @@ namespace libOpenCOR {
 
 class LIBOPENCOR_EXPORT SedTask: public SedAbstractTask
 {
-    friend class SedInstance;
+    friend class SedInstanceTask;
 
 public:
     /**
